Переделал сохранение темы на CGD
