+++ Chat/ViewModel/Theme.swift
@@ -88,65 +88,3 @@ enum ThemeName: String, CaseIterable {
         }
     }
 }
-
-class ThemeManager {
-    static let shared = ThemeManager()
-    private static let keyForStorage = "currentTheme"
-    
-    private(set) var currentThemeName: ThemeName? {
-        didSet {
-            apply()
-        }
-    }
-    
-    private init() {}
-    
-    public func load() -> Void {
-        let storedThemeAsAny = UserDefaults.standard.value(forKey: Self.keyForStorage)
-        
-        if let storedThemeAsString = storedThemeAsAny as? String,
-           let storedThemeName = ThemeName(rawValue: storedThemeAsString) {
-            currentThemeName = storedThemeName
-        } else {
-            currentThemeName = .classic
-        }
-
-    }
-    
-    public func saveTheme(with name: ThemeName) -> Void {
-        UserDefaults.standard.set(name.rawValue, forKey: Self.keyForStorage)
-        currentThemeName = name
-    }
-    
-    private func apply() -> Void {
-        guard let theme = currentThemeName?.theme else { return }
-        
-        
-        // ThemedView
-        ThemedView.appearance().backgroundColor = theme.backgroundColor
-        UILabel.appearance(whenContainedInInstancesOf: [ThemedView.self]).textColor = theme.textColor
-        UILabel.appearance(whenContainedInInstancesOf: [UserAvatarView.self, ThemedView.self]).textColor = UIColor(red: 0.21, green: 0.22, blue: 0.22, alpha: 1.00)
-       
-        // UITableView
-        UITableView.appearance().backgroundColor = theme.backgroundColor
-        
-        // NavigationBar
-        UINavigationBar.appearance().barTintColor = theme.secondBackgroundColor
-        UINavigationBar.appearance().titleTextAttributes = [.foregroundColor: theme.textColor]
-        
-        // ConversationsTableViewCell
-        ConversationsTableViewCell.appearance().backgroundColor = theme.backgroundColor
-        ConversationsTableViewCell.appearance().primaryTextColor = theme.textColor
-        ConversationsTableViewCell.appearance().onlineBackgroundColor = theme.onlineConverationCellColor
-        ConversationsTableViewCell.appearance().onlineNameTextColor = theme.onlineConverationCellTextColor
-        
-        // MessageTableViewCell
-        MessageTableViewCell.appearance().backgroundColor = theme.backgroundColor
-        MessageTableViewCell.appearance().incomeMessageCellColor = theme.incomeMessageCellColor
-        MessageTableViewCell.appearance().outcomeMessageCellColor = theme.outcomeMessageCellColor
-        MessageTableViewCell.appearance().incomeDateTextColor = theme.incomeDateTextColor
-        MessageTableViewCell.appearance().outcomeDateTextColor = theme.outcomeDateTextColor
-        MessageTableViewCell.appearance().incomeMessageTextColor = theme.incomeMessageTextColor
-        MessageTableViewCell.appearance().outcomeMessageTextColor = theme.outcomeMessageTextColor
-    }
-}
